pyproject.toml
[project]
name = "pima-diabetes-prediction"
version = "0.1.0"
description = "Production-ready diabetes prediction pipeline optimized for recall"
authors = [{name = "ML Team"}]
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "scikit-learn>=1.4.0",
    "lightgbm>=4.3.0",
    "optuna>=3.5.0",
    "mlflow>=2.10.0",
    "pydantic>=2.6.0",
    "pandera>=0.18.0",
    "evidently>=0.4.0",
    "pyyaml>=6.0.1",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "joblib>=1.3.0",
    "pytest>=8.0.0",
    "black>=24.0.0",
    "ruff>=0.2.0",
    "dvc>=3.63.0",
]

[project.optional-dependencies]
dev = [
    "marimo>=0.1.0",
    "ipykernel>=6.29.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
